endpoints/requirements.txt
flask
flask-cors
nltk
numpy
tensorflow
wikipedia
pytube
youtube.py
youtube-search-python